dedicate gesture surface view
