+++ surfaceviewrxed/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     tools:context=".ui.main.MainActivity">
 
-    <com.rosberry.sample.surfaceviewrxed.ui.main.DrivenSurfaceView
+    <com.rosberry.sample.surfaceviewrxed.ui.main.GestureSurfaceView
         android:id="@+id/surfaceView"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
